+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+* Deprecate `error::NoError` in favor of `std::convert::Infallible` ([#688])
+* Use `#[non_exhaustive]` for `error::Error`. Note this bumps the minimum supported rust version to 1.40 ([#688]).
+
+[#688]: https://github.com/amethyst/specs/pull/688
+
 # 0.16.1 (2020-02-18)
 
 * `JoinIter` now implements `Clone` when inner types are `Clone` -- usually for immutable `join()`s. ([#620])